Create longest_subsequence.cpp
You are given with an array of integers that contain numbers in random order. Write a program to find the longest possible sub sequence of consecutive numbers using the numbers from given array.
You need to return the output array which contains consecutive elements. Order of elements in the output is not important.
Best solution takes O(n) time.
If two arrays are of equal length return the array whose index of smallest element comes first.



Sample Input 1 :
13
2 12 9 16 10 5 3 20 25 11 1 8 6 


Sample Output 1 :
8 
9 
10 
11 
12